// server/src/types/services/selectBattery.type.ts
type SelectBattery_Type = {
    zendureSolarflow2400AC_N1: {
        sn: string;
        status: boolean;
        electricLevel: number;
    };
    zendureSolarflow2400AC_N2: {
        sn: string;
        status: boolean;
        electricLevel: number;
    };
}

export type { SelectBattery_Type }